chore: move dialog into its own directory
