+++ src/components/Dialog/Dialog.spec.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import Dialog from './Dialog';
+import Dialog from './';
 import { render } from '@testing-library/react';
 
 describe('Dialog component', () => {
