+++ src/components/Dialog/index.ts
@@ -1,0 +1,4 @@
+import Dialog, { SourceProps as _SourceProps } from './Dialog';
+
+export default Dialog;
+export type SourceProps = _SourceProps;
